backend/requirements.txt
Django>=3.2
djangorestframework
psycopg2-binary
qrcode
Pillow
django-cors-headers
djangorestframework-simplejwt
pandas
numpy